pyproject.toml
[project]
name = "tanda-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.8",
    "langchain-anthropic>=1.1.0",
    "langchain-openai>=1.0.3",
    "python-dotenv>=1.2.1",
]
